Split out private Client funcs, temporary ATS disable
